Bot/requirements.txt
Flask==2.3.3
python-telegram-bot==21.7
requests>=2.28.1,<3.0.0
python-dotenv>=0.20.0
cron_descriptor
pydub

